uint16_t

+++ src/thingsboard.h
@@ -146,11 +146,11 @@ class ThingsBoardSized
     // Destroys ThingsBoardSized class with network client.
     inline ~ThingsBoardSized() { }
 
-    bool setBufferSize(uint32_t size)
+    bool setBufferSize(uint16_t size)
     {
       return m_client.setBufferSize(size);
     }
-    uint32_t getBufferSize()
+    uint16_t getBufferSize()
     {
       return m_client.getBufferSize();
     }
